# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "Alert/wfgetActualAlerts": {
            "get": {
                "tags": [
                    "Alerta"
                ],
                "summary": "Get the user ID of the actual alert request",
                "description": "Request user ID for alert",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "body",
                        "description": "Alert that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "actual_alert": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "Alert/wfgetNotActualAlert": {
            "get": {
                "tags": [
                    "Alerta"
                ],
                "summary": "Get the user ID of the no actual alert request",
                "description": "Request user ID for no actual alert",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "body",
                        "description": "Alert that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Certificado"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "noactual_alert": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "Alert/Location/{alert_id}": {
            "get": {
                "tags": [
                    "Alerta"
                ],
                "summary": "Alert ID will be used to return",
                "description": "The alert ID will be used to return the url, countryName, countryCode and timestamps",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "body",
                        "description": "Alert that should be stored",
                        "required": true,
                        "schema": {
                            "$ref": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "url": {
                                    "type": "string"
                                },
                                "countryName": {
                                    "type": "string"
                                },
                                "countryCode": {
                                    "type": "string"
                                },
                                "timestamps": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Certificado": {
            "required": [
                "user_id",
                "version",
                "autor",
                "titulo",
                "contenido",
                "uuid",
                "red",
                "traza",
                "ipfs",
                "bloqueado"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "description": "user_id",
                    "type": "integer",
                    "format": "int32"
                },
                "version": {
                    "description": "version",
                    "type": "integer",
                    "format": "int32"
                },
                "autor": {
                    "description": "autor",
                    "type": "string"
                },
                "titulo": {
                    "description": "titulo",
                    "type": "string"
                },
                "contenido": {
                    "description": "contenido",
                    "type": "string"
                },
                "uuid": {
                    "description": "uuid",
                    "type": "string"
                },
                "red": {
                    "description": "red",
                    "type": "string"
                },
                "traza": {
                    "description": "traza",
                    "type": "string"
                },
                "ipfs": {
                    "description": "ipfs",
                    "type": "string"
                },
                "clave": {
                    "description": "clave",
                    "type": "string"
                },
                "bloqueado": {
                    "description": "bloqueado",
                    "type": "boolean"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}